ref(domain): manually commit and rollback the tx.
The defer approach will probably catching an error outside TCL related error and also it could not catch user's defined error directly (we have to re-assign err variable with our error).

We can isolate the transaction in a closure or just separate the data access logic in another layer (e.g. repository). But that's for the other day.